+++ src/store/queueSync.ts
@@ -1,5 +1,5 @@
+import { savePlayQueue } from '../api/subsonicPlayQueue';
 import type { Track } from './playerStoreTypes';
-import { savePlayQueue } from '../api/subsonic';
 import { getPlaybackProgressSnapshot } from './playbackProgress';
 import { usePlayerStore } from './playerStore';
 /**
